reorganize menu files
