+++ apps/expo/src/app/item/RateItem.tsx
