+++ apps/expo/src/components/index.ts
@@ -1,2 +1,3 @@
 export * from "./navigation";
 export * from "./ui";
+export * from "./menu";
