+++ apps/expo/src/components/menu/index.ts
@@ -1,0 +1,1 @@
+export * from "./Home";
